Big update to 'install/scripts' :beers::
* Move all scripts to their own directory
* Use a ini file for global configuration
* Use now a unique do.py script to control the other ones (With
* interactive menu)
* Convert everything to python except the github repos script
* Unify formatting & colors
* Add ppa for papirus icons set
* Add autopep8 pip package
* Add snap install/update functions & packages (empty for now)
* Add a json file for dotfiles to link/copy
* Add glyr & lyvi to github_repos
* Download/pull only if a new version is available for all the scripts
* Delete old scripts

+++ install/scripts/github_repos.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-# Last modification: 2016-06-29
-
-# Clone or pull the following github repositories:
+###############################
+# Modification: 2016-07-06
+#
+# Download, clone or pull github releases/repositories.
 #	- bash-git-prompt
 #	- byzanz-window
 #	- fzf
@@ -10,6 +11,9 @@
 #	- i3gaps
 #	- tidy-html5
 #	- xkblayout-state
+#	- glyr
+#	- lyvi
+###############################
 
 DEFAULT_IFS=$IFS
 IFS='
@@ -21,7 +25,7 @@ IFS='
 
 # Colors.
 white="\033[0m"
-yellow="\033[33m"
+blue="\033[34m"
 
 curr_dir=$(pwd)
 dir="$HOME/Scripts"
@@ -31,7 +35,7 @@ dir="$HOME/Scripts"
 #########
 
 Log() {
-	echo -e "\n${yellow}##### "$1" #####${white}\n"
+	echo -e "\n${blue}# "$1"${white}\n"
 }
 
 ###################
@@ -39,43 +43,62 @@ Log() {
 ###################
 
 byzanzWindow() {
-	name="byzanz-window"
+	local local name="byzanz-window"
 	Log "$name"
 	cd "$dir"
-	
-	url=$(curl -s https://api.github.com/repos/syohex/byzanz-window/releases/latest | grep 'browser_' | cut -d '"' -f 4)
-	file=$(echo "$url" | rev | cut -d "/" -f 1 | rev)
-	
-	# We remove it if it already exists
-	if [ -d "$name" ]; then
-		rm -vrf "$name"
+
+	local repo="https://api.github.com/repos/syohex/byzanz-window/releases/latest"
+	local url=$(curl -s "$repo" | grep 'browser_' | cut -d '"' -f 4)
+	local last_version=$(curl -s "$repo" | grep 'tag_name' | cut -d '"' -f 4 | tr -d "v")
+	local file=$(echo "$url" | rev | cut -d "/" -f 1 | rev)
+
+	if [ -x "/usr/local/bin/$name" ]; then
+		local current_version=$($name -v)
+	else
+		local current_version=""
+	fi
+
+	if [ "$current_version" == "$last_version" ]; then
+		echo "Already up-to-date."
+	else
+		# We remove it if it already exists
+		if [ -d "$name" ]; then
+			rm -vrf "$name"
+			echo ""
+		fi
+
+		# We download it
+		wget --no-verbose "$url" -P .
+		echo ""
+
+		mkdir -v "$name"
+		echo ""
+
+		# We extract in the new folder
+		tar -vzxf "$file" -C "$name" --strip-components=1
+		echo ""
+
+		rm -vfr "$file"
+		cd "$name"
+		echo ""
+
+		chmod -v +x "$name"
+		echo ""
+
+		sudo ln -vsf "$dir/$name/$name" "/usr/local/bin/"
 		echo ""
 	fi
-	
-	# We download it
-	wget --no-verbose "$url" -P .
-	echo ""
-	mkdir -v "$name"
-	echo ""
-	# We extract in the new folder
-	tar -vzxf "$file" -C "$name" --strip-components=1
-	echo ""
-	rm -vfr "$file"
-	cd "$name"
-	echo ""
-	chmod -v +x "$name"
-	echo ""
-	sudo ln -vsf "$dir/$name/$name" "/usr/local/bin/"
 }
 
 tidyHtml5() {
-	name="tidy-html5"
+	local name="tidy-html5"
 	Log "$name"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		# Execute & stock
+		local state=$(git pull origin master)
 		echo ""
 	else
 		git clone https://github.com/htacg/tidy-html5 "$name" &&
@@ -83,30 +106,34 @@ tidyHtml5() {
 		echo ""
 	fi
 
-	# Install requirement
-	sudo apt-get -q install -y cmake xsltproc
-	echo ""
+	if [ "$state" == "Already up-to-date." ]; then
+		echo $state
+	else
+		# Install requirements
+		sudo apt-get -q install -y cmake xsltproc
+		echo ""
 
-	# Compile
-	cd build/cmake
-	cmake ../..
-	make clean
-	make
-	echo ""
+		# Compile
+		cd build/cmake
+		cmake ../..
+		make clean
+		make
+		echo ""
 
-	# Link
-	sudo ln -vfs "$dir/$name/build/cmake/tidy" "/usr/local/bin/tidy5"
-	echo ""
+		# Link
+		sudo ln -vfs "$dir/$name/build/cmake/tidy" "/usr/local/bin/tidy5"
+		echo ""
+	fi
 }
 
 fzf() {
-	name="fzf"
+	local name="fzf"
 	Log "$name"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		local state=$(git pull origin master)
 		echo ""
 	else
 		git clone --depth 1 https://github.com/junegunn/fzf.git "$name" &&
@@ -114,21 +141,27 @@ fzf() {
 		echo ""
 	fi
 
-	./install --all
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		./install --all
+		echo ""
+	fi
 }
 
 i3gaps() {
-	name="i3gaps"
-	Log "$name"
+	local name="i3gaps"
+	local branch="gaps-next"
+	Log "$name (${branch})"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		local state=$(git pull origin "$branch")
 		echo ""
 	else
 		git clone https://github.com/Airblader/i3 "$name" &&
-			cd "$name"
+		cd "$name"
 		echo ""
 
 		# Install dependencies
@@ -136,19 +169,24 @@ i3gaps() {
 		echo ""
 	fi
 
-	# Compile & install
-	make
-	sudo make install
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		# Compile & install
+		make
+		sudo make install
+		echo ""
+	fi
 }
 
 i3blocks() {
-	name="i3blocks"
+	local name="i3blocks"
 	Log "$name"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		local state=$(git pull origin master)
 		echo ""
 	else
 		git clone https://github.com/vivien/i3blocks "$name" &&
@@ -156,19 +194,24 @@ i3blocks() {
 		echo ""
 	fi
 
-	# Compile & install
-	make clean all
-	sudo make install
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		# Compile & install
+		make clean all
+		sudo make install
+		echo ""
+	fi
 }
 
 xkblayoutState() {
-	name="xkblayout-state"
+	local name="xkblayout-state"
 	Log "$name"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		local state=$(git pull origin master)
 		echo ""
 	else
 		git clone https://github.com/nonpop/xkblayout-state "$name" &&
@@ -176,21 +219,26 @@ xkblayoutState() {
 		echo ""
 	fi
 
-	# Compile
-	make
-	echo ""
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		# Compile
+		make
+		echo ""
 
-	sudo ln -vfs "$dir/$name/xkblayout-state" "/usr/local/bin"
+		sudo ln -vfs "$dir/$name/xkblayout-state" "/usr/local/bin"
+		echo ""
+	fi
 }
 
 bashGitPrompt() {
-	name="bash-git-prompt"
+	local name="bash-git-prompt"
 	Log "$name"
 	cd "$dir"
 
 	if [ -d "$name" ]; then
 		cd "$name"
-		git pull origin master
+		local state=$(git pull origin master)
 		echo ""
 	else
 		git clone https://github.com/magicmonty/bash-git-prompt.git "$name" --depth=1 &&
@@ -198,7 +246,69 @@ bashGitPrompt() {
 		echo ""
 	fi
 
-	ln -vsf "$dir/$name/gitprompt.sh" "$HOME/.gitprompt.sh"
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		ln -vsf "$dir/$name/gitprompt.sh" "$HOME/.gitprompt.sh"
+		echo ""
+	fi
+}
+
+glyr() {
+	local name="glyr"
+	Log "$name"
+	cd "$dir"
+
+	if [ -d "$name" ]; then
+		cd "$name"
+		local state=$(git pull origin master)
+		echo ""
+	else
+		git clone https://github.com/sahib/glyr "$name" &&
+		cd "$name"
+		echo ""
+	fi
+
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+
+		# Compilation & installation
+		cmake -DCMAKE_INSTALL_PREFIX=/usr .
+		make
+		sudo make install
+		echo ""
+	fi
+}
+
+lyvi() {
+	local name="lyvi"
+	Log "$name"
+	cd "$dir"
+
+	if [ -d "$name" ]; then
+		cd "$name"
+		local state=$(git pull origin master)
+		echo ""
+	else
+		git clone https://github.com/ok100/lyvi "$name" &&
+		cd "$name"
+		echo ""
+	fi
+
+	if [ "$state" == "Already up-to-date." ]; then
+		echo "$state"
+	else
+		# N.B: GLYR IS A REQUIREMENT (CHECK PREVIOUS FUNCTION)
+
+		# Pip requirements
+		sudo -H pip3 install --upgrade cython
+		sudo -H pip3 install --upgrade -r pip_requirements.txt
+		echo ""
+
+		sudo ln -vsf "$dir/$name/lyvi.py" "/usr/local/bin/lyvi"
+		echo ""
+	fi
 }
 
 ##############
@@ -212,6 +322,8 @@ i3blocks
 i3gaps
 tidyHtml5
 xkblayoutState
+glyr
+lyvi
 
 cd "$curr_dir"
 IFS=$DEFAULT_IFS
